Massive simplification of SimpleOfx schema, to be much more permissive.
SimpleOfx is only intended to be used to load OFX files into some meaningful structure, not to validate or to save OFX files.
The only real visible effects are that in the editor, the drop-down boxes for account type and transaction type are replaced with text boxes.
When saving, we'll just be careful to format the text properly.